solving bug

+++ src/main/java/bca/Hapepedia/repo/ColorRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import bca.Hapepedia.entity.Color;
 
-public interface ColorRepo extends PagingAndSortingRepository<Color, Integer>{
+public interface ColorRepo extends PagingAndSortingRepository<Color, Long>{
     public Optional<Color> findById(Long id);
 }
